+++ .gitignore
@@ -23,6 +23,7 @@
 *.rar
 *.iml
 *.idea
+*.idea/runConfigurations
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
