+++ src/manager/DateTimeFormat.java
@@ -1,0 +1,7 @@
+package manager;
+
+import java.time.format.DateTimeFormatter;
+
+public final class DateTimeFormat {
+    public static final DateTimeFormatter DATE_TIME_FORMAT = DateTimeFormatter.ofPattern("HH:mm_dd.MM.yyyy");
+}
