+++ test/manager/TaskManagerTest.java
@@ -9,26 +9,157 @@ import task.Task;
 
 import java.time.Duration;
 import java.time.LocalDateTime;
+import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public abstract class TaskManagerTest<T extends TaskManager> {
     protected T taskManager;
+    protected Task task;
+    protected Epic epic;
+    protected Subtask subtask1;
+    protected Subtask subtask2;
 
     protected abstract T createTestManager();
 
     @BeforeEach
     void initialize() {
         taskManager = createTestManager();
+        task = new Task("Test addNewTask", "Test addNewTask description");
+        epic = new Epic("Test addNewEpic", "Test addNewEpic description");
+    }
+
+    @Test
+    void addNewTask() {
+        final int taskId = taskManager.addTask(task);
+        final Task savedTask = taskManager.getTaskByID(taskId);
+
+        assertNotNull(savedTask, "Задача не найдена.");
+        assertEquals(task, savedTask, "Задачи не совпадают.");
+        assertEquals(task.getId(), savedTask.getId(), "ID не совпадают.");
+        assertEquals(task.getName(), savedTask.getName(), "Name не совпадают.");
+        assertEquals(task.getDescription(), savedTask.getDescription(), "Description не совпадают.");
+        assertEquals(task.getStatus(), savedTask.getStatus(), "Status не совпадают.");
+
+        final List<Task> tasks = taskManager.getTaskList();
+
+        assertNotNull(tasks, "Задачи не возвращаются.");
+        assertEquals(1, tasks.size(), "Неверное количество задач.");
+        assertEquals(task, tasks.getFirst(), "Задачи не совпадают.");
+    }
+
+    @Test
+    void addNewEpic() {
+        final int epicId = taskManager.addTask(epic);
+        final Epic savedEpic = (Epic) taskManager.getTaskByID(epicId);
+
+        assertNotNull(savedEpic, "Задача не найдена.");
+        assertEquals(epic, savedEpic, "Задачи не совпадают.");
+
+        savedEpic.addSubtask(epicId);
+        final List<Integer> subtasksOfEpic = savedEpic.getListIdOfSubtasks();
+
+        assertNotNull(subtasksOfEpic, "Задачи не возвращаются.");
+        assertEquals(0, subtasksOfEpic.size(), "Неверное количество подзадач.");
+    }
+
+    @Test
+    void addNewSubtask() {
+        final int epicId = taskManager.addTask(epic);
+        subtask1 = new Subtask("Test addNewTask", "Test addNewTask description",
+                Status.NEW, epicId);
+        final int subId1 = taskManager.addTask(subtask1);
+        subtask2 = new Subtask("Test addNewTask", "Test addNewTask description",
+                Status.NEW, subId1);
+        final int subId2 = taskManager.addTask(subtask2);
+
+        final Subtask savedSub = (Subtask) taskManager.getTaskByID(subId1);
+
+        assertNotNull(savedSub, "Задача не найдена.");
+        assertEquals(subtask1, savedSub, "Задачи не совпадают.");
+
+        assertNull(taskManager.getTaskByID(subId2), "Некорректное сохранение позадачи.");
+
+        final List<Subtask> subtasks = taskManager.getSubTaskList();
+
+        assertNotNull(subtasks, "Задачи не возвращаются.");
+        assertEquals(1, subtasks.size(), "Неверное количество подзадач.");
+    }
+
+    @Test
+    void notConflictTasksInMemoryTaskManager() {
+        int notGenId = 1111111;
+        task.setId(notGenId);
+        taskManager.updateTask(task);
+        assertNull(taskManager.getTaskByID(notGenId), "Некорректное обновление позадачи.");
+
+        final List<Task> tasks = taskManager.getTaskList();
+        int savedSize = tasks.size();
+        taskManager.deleteTaskById(notGenId);
+        assertEquals(savedSize, taskManager.getTaskList().size(), "Неверное количество задач.");
+
+        final int newTaskId = taskManager.addTask(task);
+        assertNotEquals(newTaskId, notGenId, "Новый ID при добавлении не сгенерирован");
+    }
+
+    @Test
+    void deleteTaskTest() {
+        final int taskId = taskManager.addTask(task);
+        final int epicId = taskManager.addTask(epic);
+        subtask1 = new Subtask("Test addNewTask", "Test addNewTask description",
+                Status.NEW, epicId);
+        final int subId1 = taskManager.addTask(subtask1);
+        subtask2 = new Subtask("Test addNewTask", "Test addNewTask description",
+                Status.NEW, epicId);
+        final int subId2 = taskManager.addTask(subtask2);
+
+        taskManager.deleteTaskById(taskId);
+        assertNull(taskManager.getTaskByID(taskId), "Некорректное удаление задачи.");
+
+        taskManager.deleteTaskById(subId1);
+        assertNull(taskManager.getTaskByID(subId1), "Некорректное удаление подзадачи.");
+        assertFalse(epic.getListIdOfSubtasks().contains(subId1), "В Эпике есть ссылка на удаленную подзадачу.");
+
+        taskManager.deleteTaskById(epicId);
+        assertNull(taskManager.getTaskByID(epicId), "Некорректное удаление эпика.");
+        assertNull(taskManager.getTaskByID(subId2), "При удалении эпика осталась подзадача.");
+    }
+
+    @Test
+    void deleteAllTasksTest() {
+        taskManager.addTask(task);
+        final int epicId = taskManager.addTask(epic);
+        subtask1 = new Subtask("Test addNewTask", "Test addNewTask description",
+                Status.NEW, epicId);
+        taskManager.addTask(subtask1);
+        subtask2 = new Subtask("Test addNewTask", "Test addNewTask description",
+                Status.NEW, epicId);
+        taskManager.addTask(subtask2);
+
+        taskManager.deleteAllTask();
+        assertTrue(taskManager.getTaskList().isEmpty(), "Некорректное удаление задач.");
+
+        taskManager.deleteAllSubtask();
+        assertTrue(taskManager.getSubTaskList().isEmpty(), "Некорректное удаление подзадач.");
+        assertTrue(epic.getListIdOfSubtasks().isEmpty(), "В Эпике есть список удаленных подзадач.");
+
+        taskManager.deleteAllEpic();
+        assertTrue(taskManager.getEpicList().isEmpty(), "Некорректное удаление эпика.");
     }
 
     @Test
     void reduceEpicStatusIfAllSubtasksNew() {
-        Epic epic = new Epic("Test addNewEpic", "Test addNewEpic description");
         final int epicId = taskManager.addTask(epic);
-        Subtask subtask1 = new Subtask("Test addNewSubTask1", "Test addNewTask description",
+        subtask1 = new Subtask("Test addNewSubTask1", "Test addNewTask description",
                 Status.NEW, epicId);
-        Subtask subtask2 = new Subtask("Test addNewSubTask2", "Test addNewTask description",
+        subtask2 = new Subtask("Test addNewSubTask2", "Test addNewTask description",
                 Status.NEW, epicId);
         taskManager.addTask(subtask1);
         taskManager.addTask(subtask2);
@@ -37,11 +168,10 @@ public abstract class TaskManagerTest<T extends TaskManager> {
 
     @Test
     void reduceEpicStatusIfAllSubtasksDone() {
-        Epic epic = new Epic("Test addNewEpic", "Test addNewEpic description");
         final int epicId = taskManager.addTask(epic);
-        Subtask subtask1 = new Subtask("Test addNewSubTask1", "Test addNewTask description",
+        subtask1 = new Subtask("Test addNewSubTask1", "Test addNewTask description",
                 Status.DONE, epicId);
-        Subtask subtask2 = new Subtask("Test addNewSubTask2", "Test addNewTask description",
+        subtask2 = new Subtask("Test addNewSubTask2", "Test addNewTask description",
                 Status.DONE, epicId);
         taskManager.addTask(subtask1);
         taskManager.addTask(subtask2);
@@ -50,11 +180,10 @@ public abstract class TaskManagerTest<T extends TaskManager> {
 
     @Test
     void reduceEpicStatusIfAllSubtasksInProgress() {
-        Epic epic = new Epic("Test addNewEpic", "Test addNewEpic description");
         final int epicId = taskManager.addTask(epic);
-        Subtask subtask1 = new Subtask("Test addNewSubTask1", "Test addNewTask description",
+        subtask1 = new Subtask("Test addNewSubTask1", "Test addNewTask description",
                 Status.IN_PROGRESS, epicId);
-        Subtask subtask2 = new Subtask("Test addNewSubTask2", "Test addNewTask description",
+        subtask2 = new Subtask("Test addNewSubTask2", "Test addNewTask description",
                 Status.IN_PROGRESS, epicId);
         taskManager.addTask(subtask1);
         taskManager.addTask(subtask2);
@@ -63,11 +192,10 @@ public abstract class TaskManagerTest<T extends TaskManager> {
 
     @Test
     void reduceEpicStatusIfSubtasksNewAndDone() {
-        Epic epic = new Epic("Test addNewEpic", "Test addNewEpic description");
         final int epicId = taskManager.addTask(epic);
-        Subtask subtask1 = new Subtask("Test addNewSubTask1", "Test addNewTask description",
+        subtask1 = new Subtask("Test addNewSubTask1", "Test addNewTask description",
                 Status.NEW, epicId);
-        Subtask subtask2 = new Subtask("Test addNewSubTask2", "Test addNewTask description",
+        subtask2 = new Subtask("Test addNewSubTask2", "Test addNewTask description",
                 Status.DONE, epicId);
         taskManager.addTask(subtask1);
         taskManager.addTask(subtask2);
@@ -77,11 +205,10 @@ public abstract class TaskManagerTest<T extends TaskManager> {
     @Test
     void reduceEpicTime() {
         LocalDateTime start = LocalDateTime.now();
-        Epic epic = new Epic("Test addNewEpic", "Test addNewEpic description");
         final int epicId = taskManager.addTask(epic);
-        Subtask subtask1 = new Subtask("Test addNewSubTask1", "Test addNewTask description",
+        subtask1 = new Subtask("Test addNewSubTask1", "Test addNewTask description",
                 start, 30, Status.NEW, epicId);
-        Subtask subtask2 = new Subtask("Test addNewSubTask2", "Test addNewTask description",
+        subtask2 = new Subtask("Test addNewSubTask2", "Test addNewTask description",
                 start.plusHours(1), 30, Status.NEW, epicId);
         taskManager.addTask(subtask1);
         taskManager.addTask(subtask2);
