+++ test/task/EpicTest.java
@@ -4,12 +4,12 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 public class EpicTest {
-    Epic epic1 = new Epic("Test newTask1", "Test newTask1 description");
-    Epic epic2 = new Epic("Test newTask2", "Test newTask2 description");
-    int testID = 1001;
+    private final Epic epic1 = new Epic("Test newTask1", "Test newTask1 description");
+    private final Epic epic2 = new Epic("Test newTask2", "Test newTask2 description");
 
     @Test
     void epicsEqualIfSameId() {
+        int testID = 1001;
         epic1.setId(testID);
         epic2.setId(testID);
         Assertions.assertEquals(epic1, epic2, "Задачи с одним ID не равны");
